requirements.txt
anyio==3.6.1
click==8.1.3
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
filelock==3.7.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
MarkupSafe==2.1.1
platformdirs==2.5.2
pydantic==1.9.2
PyJWT==2.4.0
PyMySQL==1.0.2
python-dotenv==0.20.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
virtualenv==20.14.1
Werkzeug==2.2.2
